+++ packages/spacecat-shared-data-access/test/unit/models/site-candidate.test.js
@@ -23,6 +23,30 @@ import {
 const validData = {
   baseURL: 'https://www.example.com',
   status: 'PENDING',
+  hlxConfig: {
+    rso: {
+      owner: 'some-owner',
+      site: 'some-site',
+      ref: 'main',
+    },
+    cdnProdHost: 'www.example.com',
+    code: {
+      owner: 'some-owner',
+      repo: 'some-repo',
+      source: {
+        type: 'github',
+        url: 'https://github.com/some-owner/some-repo',
+      },
+    },
+    content: {
+      contentBusId: '1234',
+      source: {
+        type: 'onedrive',
+        url: 'https://some-owner.sharepoint.com/:f:/r/sites/SomeFolder/Shared%20Documents/some-site/www',
+      },
+    },
+    hlxVersion: 5,
+  },
 };
 
 describe('Site Candidate Model Tests', () => {
@@ -31,10 +55,20 @@ describe('Site Candidate Model Tests', () => {
       expect(() => createSiteCandidate({ ...validData, baseURL: 'invalid-url' })).to.throw('Base URL must be a valid URL');
     });
 
+    it('throws an error if hlxConfig is invalid', () => {
+      expect(() => createSiteCandidate({ ...validData, hlxConfig: '1234' })).to.throw('HLX Config must be an object: 1234');
+    });
+
     it('creates a site candidate object with valid baseURL', () => {
       const siteCandidate = createSiteCandidate({ ...validData });
       expect(siteCandidate).to.be.an('object');
       expect(siteCandidate.getBaseURL()).to.equal(validData.baseURL);
+    });
+
+    it('creates a site candidate without hlxConfig', () => {
+      const siteCandidate = createSiteCandidate({ ...validData, hlxConfig: undefined });
+      expect(siteCandidate).to.be.an('object');
+      expect(siteCandidate.getHlxConfig()).to.deep.equal({});
     });
 
     it('creates a site candidate with default updated by', () => {
@@ -52,6 +86,34 @@ describe('Site Candidate Model Tests', () => {
 
     it('does not have an id', () => {
       expect(siteCandidate.getId()).to.be.undefined;
+    });
+
+    it('updates hlxConfig correctly', () => {
+      const newHlxConfig = {
+        cdnProdHost: 'www.another-example.com',
+        code: {
+          owner: 'another-owner',
+          repo: 'another-repo',
+          source: {
+            type: 'github',
+            url: 'https://github.com/another-owner/another-repo',
+          },
+        },
+        content: {
+          contentBusId: '1234',
+          source: {
+            type: 'onedrive',
+            url: 'https://another-owner.sharepoint.com/:f:/r/sites/SomeFolder/Shared%20Documents/another-site/www',
+          },
+        },
+        hlxVersion: 5,
+      };
+
+      siteCandidate.setHlxConfig(newHlxConfig);
+      const helixConfig = siteCandidate.getHlxConfig();
+
+      expect(helixConfig).to.be.an('object');
+      expect(helixConfig).to.deep.equal(newHlxConfig);
     });
 
     it('updates site id correctly', () => {